+++ src/API.csproj
@@ -27,4 +27,8 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Trading\" />
+  </ItemGroup>
+
 </Project>
