+++ src/Trading/Requests/BuyLimitOrder.cs
@@ -1,0 +1,72 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using GridEx.API.Trading.Responses;
+
+namespace GridEx.API.Trading.Requests
+{
+	[StructLayout(LayoutKind.Sequential, Pack = 1)]
+	public readonly struct BuyLimitOrder : IHftRequest
+	{
+		public BuyLimitOrder(long requestId, double price, double volume)
+		{
+			Size = MessageSize;
+			TypeCode = RequestTypeCode.BuyLimitOrder;
+			RequestId = requestId;
+			Price = price;
+			Volume = volume;
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public unsafe int CopyTo(byte[] buffer, int offset = 0)
+		{
+			fixed (BuyLimitOrder* thisAsPointer = &this)
+			fixed (byte* target = &buffer[offset])
+			{
+				byte* source = (byte*)thisAsPointer;
+				Buffer.MemoryCopy(source, target, MessageSize, MessageSize);
+			}
+
+			return MessageSize;
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public RejectReasonCode IsValid()
+		{
+			if (Price < PriceRange.Min || Price > PriceRange.Max)
+			{
+				return RejectReasonCode.InvalidOrderPriceRange;
+			}
+
+			if (Volume < VolumeRange.Min || Volume > VolumeRange.Max)
+			{
+				return RejectReasonCode.InvalidOrderVolumeRange;
+			}
+
+			return RejectReasonCode.Ok;
+		}
+
+		public ushort Size
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get;
+		}
+
+		public RequestTypeCode TypeCode
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get;
+		}
+
+		public long RequestId
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get;
+		}
+
+		public readonly double Price;
+		public readonly double Volume;
+
+		public static readonly ushort MessageSize = Convert.ToUInt16(Marshal.SizeOf<BuyLimitOrder>());
+	}
+}
