+++ src/Trading/Responses/ConnectionTooSlow.cs
@@ -2,24 +2,24 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace GridEx.API.Responses
+namespace GridEx.API.Trading.Responses
 {
 	[StructLayout(LayoutKind.Sequential, Pack = 1)]
-	public readonly struct RequestRejected : IHftResponse
+	public readonly struct ConnectionTooSlow : IHftResponse
 	{
-		public RequestRejected(RejectReasonCode rejectCode)
+		public ConnectionTooSlow(int responseQueueSize)
 		{
 			Size = MessageSize;
-			TypeCode = ResponseTypeCode.RequestRejected;
-			RejectCode = rejectCode;
+			TypeCode = ResponseTypeCode.ConnectionTooSlow;
+			ResponseQueueSize = responseQueueSize;
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static unsafe ref readonly RequestRejected CopyFrom(byte[] array, int offset = 0)
+		public static unsafe ref readonly ConnectionTooSlow CopyFrom(byte[] array, int offset = 0)
 		{
 			fixed (byte* source = &array[offset])
 			{
-				return ref ((RequestRejected*)source)[0];
+				return ref ((ConnectionTooSlow*)source)[0];
 			}
 		}
 
@@ -35,8 +35,8 @@ namespace GridEx.API.Responses
 			get;
 		}
 
-		public readonly RejectReasonCode RejectCode;
+		public readonly int ResponseQueueSize;
 
-		public static readonly ushort MessageSize = Convert.ToUInt16(Marshal.SizeOf<RequestRejected>());
+		public static readonly ushort MessageSize = Convert.ToUInt16(Marshal.SizeOf<ConnectionTooSlow>());
 	}
 }
