+++ src/Trading/Responses/ResponseTypeCode.cs
@@ -1,4 +1,4 @@
-﻿namespace GridEx.API.Responses
+﻿namespace GridEx.API.Trading.Responses
 {
 	public enum ResponseTypeCode : byte
     {
