+++ src/Trading/Responses/UserTokenAccepted.cs
@@ -1,9 +1,9 @@
-﻿using GridEx.API.Requests;
+﻿using GridEx.API.Trading.Requests;
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace GridEx.API.Responses
+namespace GridEx.API.Trading.Responses
 {
 	[StructLayout(LayoutKind.Sequential, Pack = 1)]
 	public readonly struct UserTokenAccepted : IHftResponse
